refactor(analyzer): Split package manager dependency classes across files
Increase the overview.
